% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JSFS_stack.r
\name{get_JSFS_stack}
\alias{get_JSFS_stack}
\title{Calculate joint site frequency spectrum from two gentoype matrices, assuming loci in same order. Data can be missing. Generates a stack of spectra for sites with equal missingness profiles.}
\usage{
get_JSFS_stack(gt_SNP_1, gt_SNP_2, n_min = 1)
}
\arguments{
\item{gt_SNP_1}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}

\item{gt_SNP_2}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a stack of joint site frequency spectra
}
\description{
Calculate joint site frequency spectrum from two gentoype matrices, assuming loci in same order. Data can be missing. Generates a stack of spectra for sites with equal missingness profiles.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
