% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JSFSm.r
\name{get_JSFSm}
\alias{get_JSFSm}
\title{Calculate joint site frequency spectrum from two gentoype matrices, assuming loci in same order. Data can be missing, but must be same number missing from each locus.}
\usage{
get_JSFSm(gt_SNP_1, gt_SNP_2)
}
\arguments{
\item{gt_SNP_1}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}

\item{gt_SNP_2}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate joint site frequency spectrum from two gentoype matrices, assuming loci in same order. Data can be missing, but must be same number missing from each locus.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
