% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SFS_MAC.r
\name{get_SFS_MAC}
\alias{get_SFS_MAC}
\title{Calculate site frequency spectrum from a gentoype matrix. Data can be missing, but must be same number missing from each locus.}
\usage{
get_SFS_MAC(gt_MAC, gt_N)
}
\arguments{
\item{gt_MAC}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate site frequency spectrum from a gentoype matrix. Data can be missing, but must be same number missing from each locus.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
