% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SFS_stack.r
\name{get_SFS_stack}
\alias{get_SFS_stack}
\title{Calculate site frequency spectrum from a gentoype matrix. Data can be missing, and will return a stack of spectra for sites with different numbers of missing genotypes.}
\usage{
get_SFS_stack(gt_SNP, n_min = 1)
}
\arguments{
\item{gt_SNP}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate site frequency spectrum from a gentoype matrix. Data can be missing, and will return a stack of spectra for sites with different numbers of missing genotypes.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
