% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SFS_stack_MAC.r
\name{get_SFS_stack_MAC}
\alias{get_SFS_stack_MAC}
\title{Calculate site frequency spectrum from a gentoype matrix. Data can be missing, and will return a stack of spectra for sites with different numbers of missing genotypes.}
\usage{
get_SFS_stack_MAC(gt_MAC, gt_N)
}
\arguments{
\item{gt_MAC}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a stack of site frequency spectra
}
\description{
Calculate site frequency spectrum from a gentoype matrix. Data can be missing, and will return a stack of spectra for sites with different numbers of missing genotypes.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
