% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SFSm.r
\name{get_SFSm}
\alias{get_SFSm}
\title{Calculate site frequency spectrum from a gentoype matrix. Data can be missing, but must be same number missing from each locus.}
\usage{
get_SFSm(gt_SNP)
}
\arguments{
\item{gt_SNP}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate site frequency spectrum from a gentoype matrix. Data can be missing, but must be same number missing from each locus.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
