% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_to_pop_minor_allele_counts.r
\name{gt_to_pop_minor_allele_counts}
\alias{gt_to_pop_minor_allele_counts}
\title{Returns a genotype matrix in counts of the minor allele.}
\usage{
gt_to_pop_minor_allele_counts(gtr, pop)
}
\arguments{
\item{gt_SNP_1}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}
}
\value{
a genotype matrix with counts of minor allele.
}
\description{
Returns a genotype matrix in counts of the minor allele.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
