% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_JSFS.r
\name{project_JSFS}
\alias{project_JSFS}
\title{Performs a down-projection of the joint site frequency spectrum}
\usage{
project_JSFS(jsfs, m, renorm = FALSE)
}
\arguments{
\item{jsfs}{\itemize{
\item a joint site frequency spectrum \link{Required}
}}

\item{m}{\itemize{
\item value for down projection \link{Required}
}}
}
\value{
a down-projected joint site frequency spectrum
}
\description{
Performs a down-projection of the joint site frequency spectrum
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
