% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_JSFS_from_genotypes.r
\name{project_JSFS_from_genotypes}
\alias{project_JSFS_from_genotypes}
\title{Calculate and down project a joint site frequency spectrum from two gentoype matrices, assuming loci in same order. Data can be missing.}
\usage{
project_JSFS_from_genotypes(gt_SNP_1, gt_SNP_2, m)
}
\arguments{
\item{gt_SNP_1}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}

\item{gt_SNP_2}{\itemize{
\item genotype matrix for a set of individuals (in rows) \link{Required}
}}

\item{m}{\itemize{
\item value for down-projection \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate and down project a joint site frequency spectrum from two gentoype matrices, assuming loci in same order. Data can be missing.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
