% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_JSFS_stack_mc.r
\name{project_JSFS_stack_mc}
\alias{project_JSFS_stack_mc}
\title{Down-project a stack of joint site frequency spectra.}
\usage{
project_JSFS_stack_mc(JSFS_stack, m, ncpu = NULL)
}
\arguments{
\item{JSFS_stack}{\itemize{
\item a set of site frequency spectra, in list format.  \link{Required}
}}

\item{m}{\itemize{
\item value for down-projection \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Down-project a stack of joint site frequency spectra.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
