% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_SFS.r
\name{project_SFS}
\alias{project_SFS}
\title{Down project a site frequency spectrum using an approach outlined by https://doi.org/10.1534/genetics.166.1.351}
\usage{
project_SFS(sfs, m)
}
\arguments{
\item{sfs}{\itemize{
\item site frequency spectrum \link{Required}
}}

\item{m}{\itemize{
\item value for down-projection \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Down project a site frequency spectrum using an approach outlined by https://doi.org/10.1534/genetics.166.1.351
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
