% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_SFS_from_genotypes.r
\name{project_SFS_from_genotypes}
\alias{project_SFS_from_genotypes}
\title{Calculate and down-project a site frequency spectrum from a genotype matrix}
\usage{
project_SFS_from_genotypes(gt_SNP, m, n_min = 1)
}
\arguments{
\item{gt_SNP}{\itemize{
\item genotype matrix with samples in rows. \link{Required}
}}

\item{m}{\itemize{
\item value for down-projection \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate and down-project a site frequency spectrum from a genotype matrix
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
