% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_SFS_from_MAC.r
\name{project_SFS_from_MAC}
\alias{project_SFS_from_MAC}
\title{Calculate and down-project a site frequency spectrum from a genotype matrix}
\usage{
project_SFS_from_MAC(gt_MAC, gt_N, m)
}
\arguments{
\item{gt_MAC}{\itemize{
\item genotype matrix with samples in rows. \link{Required}
}}

\item{m}{\itemize{
\item value for down-projection \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Calculate and down-project a site frequency spectrum from a genotype matrix
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
