% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_SFS_stack.r
\name{project_SFS_stack}
\alias{project_SFS_stack}
\title{Down project a stack of site frequency spectra using an approach outlined by https://doi.org/10.1534/genetics.166.1.351}
\usage{
project_SFS_stack(SFS_stack, m)
}
\arguments{
\item{SFS_stack}{\itemize{
\item a stack of site frequecy spectra in list format \link{Required}
}}

\item{m}{\itemize{
\item value for down-projection \link{Required}
}}
}
\value{
a joint site frequency spectrum
}
\description{
Down project a stack of site frequency spectra using an approach outlined by https://doi.org/10.1534/genetics.166.1.351
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
